% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performing_Condition_Partition.R
\name{condition_partition}
\alias{condition_partition}
\title{Perform the step of performing condition partition for a set of filtered ordered pairs of condition}
\usage{
condition_partition(conditionAll, Rfilt, mfra, cfra, minC)
}
\arguments{
\item{conditionAll}{all conditions of the input expression matrix}

\item{Rfilt}{the filtered R, which R is the primarily identified columns of a corresponding submatrix in the ranking matrix}

\item{mfra}{the predefined parameter, and refer to "TransBic.R"}

\item{cfra}{the predefined parameter, and refer to "TransBic.R"}

\item{minC}{the predefined parameter, and refer to "TransBic.R"}
}
\value{
conditionList: the resulting condition partition according to the input Rfilt.
}
\description{
Perform the step of performing condition partition for a set of filtered ordered pairs of condition
}
\examples{
data(sim)
aa<-sample(col(sim),3)
bb<-sample(col(sim),2)
Rex<-as.vector(sapply(aa,function(x) paste(x,bb,sep =",")))
condition_partition(col(sim),Rex,mfra=0.8,cfra=0.75,minC=2)

}
