% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructing_Ranking_Matrix.R
\name{rankingM_constructing}
\alias{rankingM_constructing}
\title{Perform the step of constructing ranking matrix for the input gene expression matrix}
\usage{
rankingM_constructing(data)
}
\arguments{
\item{data}{the input gene expression matrix
(each row represents a gene and each column represents a condition)}
}
\value{
M_matrix: the corresponding ranking matrix
}
\description{
Perform the step of constructing ranking matrix for the input gene expression matrix
}
\examples{
data(sim)
rankingM_constructing(sim)

}
