% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransBic.R
\docType{data}
\name{sim}
\alias{sim}
\title{a simulated dataset}
\format{
A matrix with 8 rows and 8 columns:
}
\usage{
data(sim)
}
\description{
In the dataset, a bicluster with rows {1,2,3,4,5,6},columns {1,2,3,4,5,6} and
its pattern: for g={1,3,5,6}, [g,{1,2}]>[g,{3,4}]>[g,{5,6}] and for g={2,4}, [g,{1,2}]<[g,{3,4}]<[g,{5,6}]is implanted.
}
\examples{
#data(sim)
#View(sim)
}
\keyword{datasets}
