% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransBic.R
\name{TransBic}
\alias{TransBic}
\title{Perform the TransBic program}
\usage{
TransBic(
  data,
  minC = 2,
  minR = 2,
  e0 = 0.1,
  tfra = 0.75,
  tNum = 5,
  cfra = 0.75,
  mfra = 0.8,
  sfra = 1
)
}
\arguments{
\item{data}{normalized gene expression matrix(each column represents a condition
and each row represents a gene)}

\item{minC}{minimum number of conditions in each condition cluster from resulting condition partitions(int)}

\item{minR}{minimum number of rows of resulting biclusters(int)}

\item{e0}{error rate of the identified rows in the corresponding submatrix in the ranking matrix([0,1])}

\item{tfra}{consistency level of the identified columns of the corresponding submatrix in this iteration([0,1])
compared to the last iteration([0,1])}

\item{tNum}{number of consecutive iterations for keeping the above consistence(int)}

\item{cfra}{level to evaluate whether a selected condition is from a new condition cluster from the same partition with the
identified condition clusters ([0,1])}

\item{mfra}{level to evaluate whether two conditions are from the same condition cluster([0,1])}

\item{sfra}{fraction of the identified columns to preserve as the columns of the corresponding submatrix([0,1])}
}
\value{
Biclusters
Taken a bicluster as an example, its output format is shown as follows:
line 1_rank order of biclusters:        Bic:number
line 2_rows:                            genes_1/2
line 3_columns:                         condition partition for a identified condition subset
       (Suppose the condition partition is C1,C2,C3,...,CT from the top down,and then
       the row g_1 has [g,C1]> [g,C2]>[g,C3]>⋯>[g,CT], where [g,C] denote all expression values under condition cluster C for gene g.
       At the same time,  the row g_2 has [g,C1]< [g,C2]<[g,C3]<⋯<[g,CT])
}
\description{
Perform the TransBic program
}
\examples{
data(sim)
TransBic(sim,tNum=3,sfra = 0.9)
}
